made ViewCustomizable use an integer size value instead of qreal

+++ interfaces/nodeshapeable.cpp
@@ -4,7 +4,7 @@
 #include <QGraphicsSceneMouseEvent>
 
 
-void NodeShapeable::setElementSize(qreal newSize) {
+void NodeShapeable::setElementSize(int newSize) {
     ViewCustomizable::setElementSize(newSize);
     this->drawShape();
 }
